Добавляет модульность в проект. Вносит изменения в по прошлой домашней работе в соответствии с комментариями. Убирает файл functions.js
