+++ js/data.js
@@ -1,0 +1,50 @@
+const dataArray = () => {
+
+  const arrayMessages = [
+    'Всё отлично!',
+    'В целом всё неплохо. Но не всё.',
+    'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+    'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+    'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+    'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+  ];
+
+  const arrayNames = [
+    'Андрей',
+    'Женя',
+    'Настя',
+    'Паша',
+    'Ксюша',
+    'master225',
+    'Сергей',
+    'Дима',
+    'Олег',
+    'criti78',
+    'Антон',
+    'Денис',
+    'Карина',
+    'Полина',
+    'Маша',
+    'Саша',
+  ];
+
+  const arrayDescriptions = [
+    'Солнечное утро',
+    'Прекрасный завтрак',
+    'Такое не должно повториться!',
+    'Мой день рождения',
+    'Пляж',
+    'Горы',
+    '#СолнечнаяПогода',
+    'Помыл машину',
+    'Подсолнухи',
+    'Мы вместе!',
+    'Каждый день - новая страна!',
+    'Ночной город - прекрасен',
+  ];
+
+  return {arrayMessages, arrayNames, arrayDescriptions};
+
+};
+
+export { dataArray };
